Expanded DownloadFile

+++ src/com/torrent/file/FileManager.java
@@ -15,11 +15,18 @@ public class FileManager {
 		}
 	}
 
-	public void writeToRoot(DownloadFile file) throws Exception {
-		FileOutputStream out = new FileOutputStream(mRoot);
-		out.write(file.getBytes());
-		out.flush();
-		out.close();
+	public void writeFile(DownloadFile file) throws Exception {
+		// If there is only one file (file.getPath() == null), then mRoot is the output file.
+		// Otherwise, mRoot is a folder and this file should be under mRoot
+		File outFile = (file.getPath() == null) ? (mRoot) : (new File(mRoot, file.getPath()));
+
+		// Double check the file is downloaded
+		if(file.isFileDownloaded()){
+			FileOutputStream out = new FileOutputStream(mRoot);
+			out.write(file.getBytes());
+			out.flush();
+			out.close();
+		}
 	}
 
 }
